requirements.txt
numpy
Pillow
keras==2.3.1
tensorflow==1.14.0
protobuf==3.20.0
h5py==2.10.0
streamlit